QtWidgets: extracted OrdinalBaseWidget
